Bugfix: Multiple flexigrids on same web page
When flexigrid is used multiple times on the same web page and at least one
instance of flexigrid has 'usepager: false' the status line at the bottom of
those flexigrid tables that have 'usepager: true' is updated incorrectly.

The reason is that flexigrid searches the DOM for the right element
to update starting from this.pDiv, but this.pDiv is not set if there is no
status line. In that case the search for the DOM element starts at the
root element, so it will find all other flexigrid instances.

This commit fixes this issue. It also stores the relevant elements instead of
searching through the DOM each time, which should make things slightly faster.